requirements.txt
crewai[tools]==1.10.1
exa-py==2.7.0
pydantic==2.11.9
python-dotenv==1.1.1
pandas==3.0.1
matplotlib==3.10.8
seaborn==0.13.2
